restructure auth
